+++ src/shared/components/Alert/index.js
@@ -1,306 +1,117 @@
-import React, { Component } from 'react';
+import { Component } from 'react';
 import { PropTypes } from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import { Slide, Collapse, Paper, Grid, ClickAwayListener } from '@material-ui/core';
-import { keys as settingsKeys } from '../../settings';
+import { AlertsContext } from '../../context/AlertsProvider';
+import { createRcaSettings } from '../..';
 import types from '../../utils/types';
 
-const styles = {
-  componentWrapper: {
-    width: '100%',
-    marginBottom: 15,
-    display: 'block',
-  },
-};
+class Alert extends Component {
+  state = {};
 
-class Alert extends Component {  
-  static determineValue = (globalValue, typeValue) => {
-    return (typeValue || typeValue === false) ? typeValue : globalValue;
-  };
+  componentDidUpdate(prevProps) {
+    const { info, success, warning, error } = this.context;
+    const { open, onClose, type, header, message } = this.props;
 
-  static settingValue = (key, type, settings) => {
-    return Alert.determineValue(settings[key], settings[type][key]);
-  };
+    const settings = createRcaSettings(
+      this.extractInstanceSettingsFromProps(),
+      this.context.globalSettings()
+    );
 
-  state = {
-    timedOut: false,
-    collapse: false,
-  };
-
-  componentDidMount() {
-    const { type, settings } = this.props;
-    const timeout = Alert.settingValue(
-      settingsKeys.timeout, type, settings);
-    if (timeout) {
-      this.timeoutFunction = setTimeout(() => {
-        this.onClose();
-      }, timeout);
+    if (open && !prevProps.open) {
+      switch(type) {
+        case types.info:
+          info(header, message, onClose, settings);
+          break;
+        case types.success:
+          success(header, message, onClose, settings);
+          break;
+        case types.warning:
+          warning(header, message, onClose, settings);
+          break;
+        case types.error:
+          error(header, message, onClose, settings);
+          break;
+        default:
+          info(header, message, onClose, settings);
+      }
     }
-  };
+  }
 
-  onClose = () => {
-    clearTimeout(this.timeoutFunction);
-    this.setState({ timedOut: true, timeoutFunction: null });
-  };
+  extractInstanceSettingsFromProps = () => {
+    const { type } = this.props;
+    const instanceSettings = { [type]: {} };
 
-  onSlideExited = () => {
-    this.setState({ collapse: true });
-  };
-
-  onCollapsed = () => {
-    this.props.onClose();
-  };
-
-  wrapperStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      background: theme[type].body.background,
-    };
-  };
-
-  bodyWrapperStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      background: theme[type].body.background,
-      color: theme[type].body.color,
-    };
-  };
-
-  bodyContentStyles = () => {
-    const { settings } = this.props;
-    const paddingLeft = this.showAdornment(settings) ? 0 : 16;
-    return {
-      paddingLeft,
-    };
-  };
-
-  headerStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      color: theme[type].body.headerColor,
-    };
-  };
-
-  messageStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      color: theme[type].body.messageColor,
-    };
-  };
-
-  bodyContent = (header, message) => {
-    const { type, theme } = this.props;
-    const content = theme[type].body.content;
-    return content(
-      theme[type].body.header(header, this.headerStyles()),
-      theme[type].body.message(message, this.messageStyles()),
-      this.bodyContentStyles());
-  };
-
-  bodyWrapper = (header, message) => {
-    const { type, theme } = this.props;
-    const wrapper = theme[type].body.wrapper;
-    return wrapper(
-      this.bodyContent(header, message),
-      this.bodyWrapperStyles());
-  };
-
-  adornmentStyles = () => {
-    const { type, theme, settings } = this.props;
-    const styles = {
-      background: theme[type].adornment.background,
-      color: theme[type].adornment.color,
-    };
-    if (this.showProgress(settings)) {
-      styles.borderBottomLeftRadius = 0;
+    if (this.props.timeout !== undefined) {
+      instanceSettings.timeout = this.props.timeout;
+      instanceSettings[type].timeout = this.props.timeout;
     }
-    return styles;
-  };
-
-  adornmentWrapper = () => {
-    const { type, theme } = this.props;
-    const wrapper = theme[type].adornment.wrapper;
-    return wrapper(theme[type].adornment.icon, this.adornmentStyles());
-  };
-  
-  actionStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      background: theme[type].action.background,
-      color: theme[type].action.color,
-      fontWeight: theme[type].action.fontWeight,
-    };
-  };
-
-  actionWrapper = (settings) => {
-    const { type, theme } = this.props;
-    const wrapper = theme[type].action.wrapper;
-    const text = Alert.settingValue(
-      settingsKeys.actionText, type, settings);
-    const onClick = Alert.settingValue(
-      settingsKeys.actionClickListener, type, settings);
-    return wrapper(
-      theme[type].action.button(text, onClick),
-      this.actionStyles());
-  };
-
-  closeStyles = () => {
-    const { type, theme } = this.props;
-    return {
-      background: theme[type].close.background,
-      color: theme[type].close.color,
-    };
-  };
-
-  closeWrapper = (onClick) => {
-    const { type, theme } = this.props;
-    const wrapper = theme[type].close.wrapper;
-    const icon = theme[type].close.icon;
-    return wrapper(
-      theme[type].close.button(icon, onClick),
-      this.closeStyles());
-  };
-
-  progressWrapper = (settings) => {
-    const { type, theme } = this.props;
-    const wrapper = theme[type].progress.wrapper;
-    const timeout = Alert.settingValue(
-      settingsKeys.timeout, type, settings);
-    return wrapper(
-      theme[type].progress.containerColor,
-      theme[type].progress.bar(
-        theme[type].progress.barColor, timeout));
-  };
-
-  mouseEventType = (settings) => {
-    const { type } = this.props;
-    const enableClickAwayListener = Alert.settingValue(
-      settingsKeys.enableClickAwayListener, type, settings);
-    return enableClickAwayListener ? 'onMouseUp' : false;
-  };
-
-  touchEventType = (settings) => {
-    const { type } = this.props;
-    const enableClickAwayListener = Alert.settingValue(
-      settingsKeys.enableClickAwayListener, type, settings);
-    return enableClickAwayListener ? 'onTouchEnd' : false;
-  };
-
-  onClickAway = () => {
-    this.onClose();
-  };
-
-  bodyColumns = (settings) => {
-    let bodyCols = 6;
-    if (!this.adornmentColumns(settings)) {
-      bodyCols += 2;
+    if (this.props.showAdornment !== undefined) {
+      instanceSettings.showAdornment = this.props.showAdornment;
+      instanceSettings[type].showAdornment = this.props.showAdornment;
     }
-    if (!this.actionButtonColumns(settings)) {
-      bodyCols += 3;
+    if (this.props.showActionButton !== undefined) {
+      instanceSettings.showActionButton = this.props.showActionButton;
+      instanceSettings[type].showActionButton = this.props.showActionButton;
     }
-    if (!this.closeButtonColumns(settings)) {
-      bodyCols += 1;
+    if (this.props.showCloseButton !== undefined) {
+      instanceSettings.showCloseButton = this.props.showCloseButton;
+      instanceSettings[type].showCloseButton = this.props.showCloseButton;
     }
-    return bodyCols;
-  };
+    if (this.props.showProgressBar !== undefined) {
+      instanceSettings.showProgressBar = this.props.showProgressBar;
+      instanceSettings[type].showProgressBar = this.props.showProgressBar;
+    }
+    if (this.props.enableClickAwayListener !== undefined) {
+      instanceSettings.enableClickAwayListener = this.props.enableClickAwayListener;
+      instanceSettings[type].enableClickAwayListener = this.props.enableClickAwayListener;
+    }
+    if (this.props.actionText !== undefined) {
+      instanceSettings.actionText = this.props.actionText;
+      instanceSettings[type].actionText = this.props.actionText;
+    }
+    if (this.props.actionClickListener !== undefined) {
+      instanceSettings.actionClickListener = this.props.actionClickListener;
+      instanceSettings[type].actionClickListener = this.props.actionClickListener;
+    }
 
-  adornmentColumns = (settings) => {
-    return this.showAdornment(settings) ? 2 : false;
-  };
-
-  actionButtonColumns = (settings) => {
-    return this.showAction(settings) ? 3 : false;
-  };
-
-  closeButtonColumns = (settings) => {
-    return this.showClose(settings) ? 1 : false;
-  };
-
-  showAdornment = (settings) => {
-    const { type } = this.props;
-    return Alert.settingValue(
-      settingsKeys.showAdornment, type, settings);
-  };
-
-  showAction = (settings) => {
-    const { type } = this.props;
-    return Alert.settingValue(settingsKeys.showActionButton, type, settings) &&
-      Alert.settingValue(settingsKeys.actionText, type, settings) &&
-      Alert.settingValue(settingsKeys.actionClickListener, type, settings);
-  };
-
-  showClose = (settings) => {
-    const { type } = this.props;
-    return Alert.settingValue(
-      settingsKeys.showCloseButton, type, settings);
-  };
-
-  showProgress = (settings) => {
-    const { type } = this.props;
-    return Alert.settingValue(
-      settingsKeys.showProgressBar, type, settings);
+    return instanceSettings;
   };
 
   render() {
-    const { timedOut, collapse } = this.state;
-    const { classes, header, message, settings } = this.props;
-    return (
-      <ClickAwayListener
-        mouseEvent={this.mouseEventType(settings)}
-        touchEvent={this.touchEventType(settings)}
-        onClickAway={this.onClickAway}
-      >
-        <Collapse in={!collapse} onExited={this.onCollapsed} unmountOnExit>
-          <Slide direction={'left'} in={!timedOut} onExited={this.onSlideExited}>
-            <Paper className={classes.componentWrapper} style={this.wrapperStyles()} elevation={4}>
-              <Grid container spacing={0}>
-                <Grid item xs={this.adornmentColumns(settings)}>
-                  {this.showAdornment(settings) && 
-                    this.adornmentWrapper()
-                  }
-                </Grid>
-                <Grid item xs={this.bodyColumns(settings)}>
-                  {// always show body
-                    this.bodyWrapper(header, message)
-                  }
-                </Grid>
-                <Grid item xs={this.actionButtonColumns(settings)}>
-                  {this.showAction(settings) &&
-                    this.actionWrapper(settings)
-                  }
-                </Grid>
-                <Grid item xs={this.closeButtonColumns(settings)}>
-                  {this.showClose(settings) &&
-                    this.closeWrapper(this.onClose)
-                  }
-                </Grid>
-              </Grid>
-              {this.showProgress(settings) &&
-                this.progressWrapper(settings)
-              }
-            </Paper>
-          </Slide>
-        </Collapse>
-      </ClickAwayListener>
-    );
+    return null;
   }
 };
+Alert.contextType = AlertsContext;
 
 Alert.propTypes = {
-  id: PropTypes.number.isRequired,
+  open: PropTypes.bool,
   onClose: PropTypes.func,
+  type: PropTypes.oneOf([types.info, types.success, types.warning, types.error]),
   header: PropTypes.string,
   message: PropTypes.string,
-  type: PropTypes.oneOf([types.info, types.success, types.warning, types.error]),
-  theme: PropTypes.object.isRequired,
-  settings: PropTypes.object.isRequired,
+  timeout: PropTypes.number,
+  showAdornment: PropTypes.bool,
+  showActionButton: PropTypes.bool,
+  showCloseButton: PropTypes.bool,
+  showProgressBar: PropTypes.bool,
+  enableClickAwayListener: PropTypes.bool,
+  actionText: PropTypes.string,
+  actionClickListener: PropTypes.func,
 };
 
 Alert.defaultProps = {
+  open: false,
   onClose: () => {},
-  header: null,
-  message: '',
   type: types.info,
+  header: null,
+  message: null,
+  timeout: undefined,
+  showAdornment: undefined,
+  showActionButton: undefined,
+  showCloseButton: undefined,
+  showProgressBar: undefined,
+  enableClickAwayListener: undefined,
+  actionText: undefined,
+  actionClickListener: undefined,
 };
 
-export default withStyles(styles)(Alert);
+export default Alert;